requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyautogen>=0.2.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API and data handling
requests>=2.30.0
yfinance>=0.2.30
alpha_vantage>=2.3.1

# Technical analysis
ta>=0.10.0
plotly>=5.14.0

# Email and reporting
jinja2>=3.1.0
pydantic>=2.0.0

# Auth and security
itsdangerous>=2.1.0